Add files via upload
This part of the assignment is based on this scenario: Assume a traveler “t” plans to visit some cities in the network given in Fig. 3. As seen in the figure, there is one-way transportation between any of two cities (directions are represented by arrows). Starting from a city, say x, t can only visit the cities whose maximum distance to x is 3. For instance, if t starts from ’C’ then the reachable city list is {’E’, ’G’, ’H’, ’A’, ’K’, ’N’, ’O’, ’M’, ’I’}; for x=’J’ reachable city list is {’K’, ’N’, ’O’}.


As in the previous part, your implementation should first read the file, (commandsP2.txt), consisting of starting cities then lists reachable cities The implementation investigates possible cities based on the descriptions above and it displays the reachable cities for the cities given in the input file.